% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fitted.logitreg}
\alias{fitted.logitreg}
\title{Get the fitted values a logitreg model.}
\usage{
\method{fitted}{logitreg}(object)
}
\arguments{
\item{object}{logitreg. An object of class logitreg.}
}
\description{
Get the fitted values a logitreg model.
}
\author{
Janek Thomas, Philipp Roesch
}

