% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.logitreg}
\alias{plot.logitreg}
\title{Plot a ROC curve for a logitreg model.}
\usage{
plot.logitreg(object, ...)
}
\arguments{
\item{object}{logitreg. An object of class logitreg.}

\item{...}{Additional parameters passed to the plot function, see
\code{\link[ROCR]{plot.performance}} for details.}
}
\description{
Plot a ROC (reciever operating characteristic) curve for a logitreg model.
See \code{\link[ROCR]{plot.performance}} for further information.
}
\author{
Janek Thomas, Philipp Roesch
}

