% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict.logitreg}
\alias{predict.logitreg}
\title{Predict values of a logitreg object}
\usage{
\method{predict}{logitreg}(object, newdata)
}
\arguments{
\item{object}{logitreg. An object of class logitreg.}

\item{newdata}{data.frame. Optional, additional data to predict. Has to have
the same dimensions as the data of the logitreg object. Tries to convert the
object to a data.frame if it is not already one.}
}
\value{
Vector of predicted probabilities.
}
\description{
Predict values of a logitreg object or for newdata with the same dimensions as
the original data.
}
\author{
Janek Thomas, Philipp Roesch
}

